% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry.R
\name{arrange_entry}
\alias{arrange_entry}
\title{Arrange an entry for consistency}
\usage{
arrange_entry(entry, challenge = "ilinet")
}
\arguments{
\item{entry}{A data.frame}

\item{challenge}{one of "ilinet", "hosp", or "state_ili", indicating which
forecasting challenge the entry is for}
}
\value{
An arranged data.frame
}
\description{
Arrange an entry for consistency
}
\keyword{internal}
