% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_entry.R
\name{construct_filename}
\alias{construct_filename}
\title{Construct the file name}
\usage{
construct_filename(team_name, week)
}
\arguments{
\item{team_name}{A character string indicating the team name}

\item{week}{A scalar numeric indicating the week}
}
\value{
A character string for the filename
}
\description{
Construct the file name
}
\keyword{internal}
