% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasonal.R
\name{create_onset}
\alias{create_onset}
\title{Creates observed truth for seasonal onset}
\usage{
create_onset(weekILI, region, year)
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values (default NULL). Must contain columns
location, week, and wILI.}

\item{region}{A character string specifying the target location - either US National 
or one of HHS Region 1-10}

\item{year}{Calendar year during which the flu season of interest begins. 
For the 2015/2016 flu season, \code{year = 2015}}
}
\value{
A data.frame with columns location, target, and bin_start_incl
}
\description{
Determines observed true values for onset week
}
\keyword{internal}

