% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasonal.R
\name{create_seasonal}
\alias{create_seasonal}
\title{Creates observed truth for seasonal targets}
\usage{
create_seasonal(weekILI, location, year, challenge = "ilinet")
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values (default NULL). 
For \code{challege = "ilinet"} or \code{challenge = "state_ili"}, must
contain columns location, week, and wILI. For \code{challenge = "hospital"},
must contain columns age_grp, week, and weeklyrate.}

\item{location}{A character string specifying the target location or age group}

\item{year}{Calendar year during which the flu season of interest begins. 
For the 2015/2016 flu season, \code{year = 2015}}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating which
challenge the submission is for (default \code{"ilinet"})}
}
\value{
A data.frame with columns location, target, and bin_start_incl
}
\description{
Determines observed true values for onset week, seasonal peak, and seasonal
peak percentage
}
\examples{
season_targets <- create_seasonal(valid_ILI, "US National")
season_targets <- create_seasonal(valid_ILI, "HHS Region 4")
  
}
