% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_truth.R
\name{create_truth}
\alias{create_truth}
\title{Creates observed truth to score entry against}
\usage{
create_truth(fluview = TRUE, weekILI = NULL)
}
\arguments{
\item{fluview}{A logical value (default TRUE) indicating whether to download ILINet from 
Fluview}

\item{weekILI}{A data.frame of weighted ILI values (default NULL). Must contain columns
location, week, and wILI. Must be NULL when downloading data using 
fluview == TRUE. Required if fluview == FALSE.}
}
\value{
A data.frame with columns location, target, and bin_start_incl
}
\description{
Determines observed true values for each target
}

