% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_week.R
\name{create_week}
\alias{create_week}
\title{Creates observed truth for 1-4 week ahead prediction}
\usage{
create_week(weekILI, start_wk, end_wk, challenge = "ilinet")
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values (default \code{NULL}). Must 
contain columns location, week, and wILI.}

\item{start_wk}{The first MMWR week used for creating forecasts}

\item{end_wk}{The last MMWR week used for creating forecasts}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating which
challenge the submission is for (default \code{"ilinet"})}
}
\value{
A data.frame with columns location, target, and bin_start_incl
}
\description{
Determines observed true values for 1 to 4 week ahead weighted ILINet predictions
}
\examples{
week_targets <- create_week(valid_ILI, 42, 18)

}
