% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_truth.R
\name{expand_percent}
\alias{expand_percent}
\title{Expands truth for percents}
\usage{
expand_percent(truth, expand)
}
\arguments{
\item{truth}{A data.frame with a single row and, at minimum, columns unit
and value}

\item{expand}{The number of percent bins to expand around the observed truth.}
}
\value{
A data.frame with the same columns as `truth` but additional rows
around the observed value.
}
\description{
Expands truth for percents
}
\keyword{internal}

