% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_truth.R
\name{expand_truth}
\alias{expand_truth}
\title{Creates a range of bins to score}
\usage{
expand_truth(truth, week_expand = 1, percent_expand = 5)
}
\arguments{
\item{truth}{A data.frame including columns unit and bin_start_incl}

\item{week_expand}{A scalar numeric (default 1) indicating the number of
bins to include around the observed value for week targets.}

\item{percent_expand}{A scalar numeric (default 5) indicating the number
of bins to include around the observed value for percent targets.}
}
\value{
A data.frame with expanded truth
}
\description{
Expands the observed truth across a range of bins
}

