% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_truth.R
\name{expand_week}
\alias{expand_week}
\title{Expands truth for weeks}
\usage{
expand_week(truth, expand)
}
\arguments{
\item{truth}{A data.frame with a single row and, at minimum, columns unit
and value}

\item{expand}{The number of weeks to expand around the observed truth.}
}
\value{
A data.frame with the same columns as `truth` but additional rows
around the observed value.
}
\description{
The week target is peculiar because MMWR weeks run from 1-52 (or 53
depending on the year) and week 1 is sequentially, immediately after week 52
(or 53).
}
\keyword{internal}

