% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_entry_hosp_score}
\alias{full_entry_hosp_score}
\title{A full valid hospitalization forecast entry containing forecast week for 
scoring}
\format{A dta frame with 6466 rows and 8 variables:
\describe{
  \item{age_grp}{character vector of age group}
  \item{target}{character vector of targets}
  \item{type}{character vector with elements "Bin" or "Point"}
  \item{unit}{character vector with elements "percent" or "week"}
  \item{bin_start_incl}{character vector with forecast bin start value (inclusive)}
  \item{bin_end_notincl}{character vector the forecast bin end value}
  \item{value}{numeric vector with probabilities or point forecasts}
  \item{forecast_week}{numeric vector of MMWR week forecasted value is based on}
}}
\usage{
full_entry_hosp_score
}
\description{
A data.frame containing a scorable valid entry for hospitalization 
forecasting
}
\keyword{datasets}
