% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_entry_state}
\alias{full_entry_state}
\title{A full valid entry for state ILI forecasting}
\format{A dta frame with 24290 rows and 7 variables:
\describe{
  \item{location}{character vector of location}
  \item{target}{character vector of targets}
  \item{type}{character vector with elements "Bin" or "Point"}
  \item{unit}{character vector with elements "percent" or "week"}
  \item{bin_start_incl}{character vector with forecast bin start value (inclusive)}
  \item{bin_end_notincl}{character vector the forecast bin end value}
  \item{value}{numeric vector with probabilities or point forecasts}
}}
\usage{
full_entry_state
}
\description{
A data.frame containing a valid entry for state ILI forecasting
}
\keyword{datasets}
