% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_point_forecasts.R
\name{generate_point_forecast}
\alias{generate_point_forecast}
\title{Generate point forecasts from probabilistic forecasts}
\usage{
generate_point_forecast(d)
}
\arguments{
\item{d}{A data.frame with columns `bin_start_incl` and `value`}
}
\value{
A data.frame with columns `type` and `value`
}
\description{
The point forecast is taken to be the expected value of the
probabilistic forecasts.
}
\seealso{
generate_point_forecasts
}
\keyword{internal}

