% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_point_forecasts.R
\name{generate_point_forecasts}
\alias{generate_point_forecasts}
\title{Generate point forecasts for all locations and targets}
\usage{
generate_point_forecasts(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
A data.frame of point forecasts for all locations and targets.
}
\description{
Generate point forecasts for all locations and targets
}
\seealso{
\code{\link{generate_point_forecast}}, \code{\link{verify_entry}}
}

