% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_entry.R
\name{get_week}
\alias{get_week}
\title{Get the surveillance week}
\usage{
get_week(date)
}
\arguments{
\item{date}{A Date object or one that can be coerced to a Date object}
}
\value{
The numeric week number
}
\description{
Get the surveillance week
}
\seealso{
\code{\link{write_entry}}, \code{\link{as.Date}}
}
\keyword{internal}

