% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_truth_1617}
\alias{hosp_truth_1617}
\title{Observed FluSurv-Net hospitalizations truth for 2016/2017 influenza season}
\format{A dta frame with 480 rows and 4 variables:
\describe{
  \item{target}{character vector of targets}
  \item{age_grp}{character vector of age groups}
  \item{forecast_week}{numeric vector of MMWR week forecasted value is based on}
  \item{bin_start_incl}{numeric vector of observed truth}
}}
\usage{
hosp_truth_1617
}
\description{
A data.frame containing the observed national hospitalization forecasting 
truth for the 2016/2017 influenza season
}
\keyword{datasets}
