% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimal_entry}
\alias{minimal_entry}
\title{A minimal valid entry}
\format{A dta frame with 729 rows and 7 variables:
\describe{
  \item{location}{character vector of locations}
  \item{target}{character vector of targets}
  \item{type}{character vector with elements "Bin" or "Point"}
  \item{unit}{character vector with elements "percent" or "week"}
  \item{bin_start_incl}{character vector with forecast bin start value (inclusive)}
  \item{bin_end_notincl}{character vector the forecast bin end value}
  \item{value}{numeric vector with probabilities or point forecasts}
}}
\usage{
minimal_entry
}
\description{
A data.frame containing a valid entry
}
\keyword{datasets}

