% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_probs.R
\name{normalize_probs}
\alias{normalize_probs}
\title{Normalize probabilistic predictions}
\usage{
normalize_probs(entry)
}
\arguments{
\item{entry}{An entry data.frame with columns location, target, type, unit, 
bin_start_incl, bin_end_incl, and value.}
}
\value{
An entry data.frame with normalized probabilities
}
\description{
Checks an entry for validity and, if valid, normalizes all probabilistic
predictions so that the cumulative probability for a target sums to 1.
}
\seealso{
\code{\link{verify_entry}}
}

