% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{past_baselines}
\alias{past_baselines}
\title{Past ILINet baselines for all locations}
\format{A data frame with 110 rows and 3 variables:
\describe{
  \item{location}{character vector of location}
  \item{year}{numeric vector with the start year of the flu season}
  \item{value}{numeric vector with the ILINet baseline value}
}}
\usage{
past_baselines
}
\description{
A data.frame containing a past ILINet baselines
}
\keyword{datasets}
