% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdc.R
\name{plot_cdc}
\alias{plot_cdc}
\title{Plots CDC Data}
\usage{
plot_cdc(region, years = NA)
}
\arguments{
\item{region}{Specifies the region to be plotted.}

\item{years}{Specifies years to retrieve data for (i.e. 2014 for CDC flu season 2014-2015). Default value is the current year and all years values should be > 1997}
}
\description{
This function allows you to plot current and past CDC data.
}
\examples{
plotcdcdata()
}
\keyword{CDC}
\keyword{FluView}
