% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plots a CDC FluView Submission}
\usage{
plot_forecast(dat, ilimx, years = NA, pdfloc = NULL)
}
\arguments{
\item{dat}{Expects a CDC fluview submission (see \code{FluSight} package for a minimal submission)}

\item{ilimx}{Numeric. Max level of ILI percentage to be plotted}

\item{years}{Year to retrieve data for}

\item{pdfloc}{Loction to print out to pdf. If NULL, will show plots in house.}
}
\description{
This function plots a quick grid of plots to view a CDC fluview submission.
}
\examples{
plot_forecast(dat=full_entry, ilimx=10, years=2016)
}
\keyword{grid}
\keyword{plots}
