% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_onset.R
\name{plot_onset}
\alias{plot_onset}
\title{Plots Onset Forecasts}
\usage{
plot_onset(dat, region)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{region}{Specifies the region to be plotted}
}
\description{
This function allows you to plot Onset Predictions
}
\examples{
plotOnset()
}
\keyword{Onset}
\keyword{Plot}
\keyword{Prediction}
