% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peakweek.R
\name{plot_peakweek}
\alias{plot_peakweek}
\title{Plots Peak Week Forecasts}
\usage{
plot_peakweek(dat, region)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{region}{Specifies the region to be plotted}
}
\description{
This function allows you to plot Peak Week Predictions
}
\keyword{Peak}
\keyword{Plot}
\keyword{Prediction}
\keyword{Week}
