% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weekahead.R
\name{plot_weekahead}
\alias{plot_weekahead}
\title{Plots Week Ahead Forecasts}
\usage{
plot_weekahead(dat, region, wk = 1, ilimax, years = NA)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{region}{Specifies the region to be plotted}

\item{wk}{Numeric. How many weeks ahead to plot. Defaults to 1.}

\item{ilimax}{Numeric. Max level of ILI percentage to be plotted}

\item{plot_current}{Logical. Whether or not to plot current ILI level.}
}
\description{
This function allows you to plot Weak Ahead Predictions
}
\examples{
plotWeekAhead()
}
\keyword{Ahead}
\keyword{Plots}
\keyword{Prediction}
\keyword{Week}
