% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry.R
\name{read_entry}
\alias{read_entry}
\title{Read in a csv entry file}
\usage{
read_entry(file)
}
\arguments{
\item{file}{A csv file path}
}
\value{
An arranged data.frame
}
\description{
This function reads in the csv file and arranges it for consistency.
}
