% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_invalid.R
\name{remove_invalid}
\alias{remove_invalid}
\title{Remove invalid predicted probabilities for scoring}
\usage{
remove_invalid(entry)
}
\arguments{
\item{entry}{An entry data.frame with columns location, target, type, unit, 
bin_start_incl, bin_end_incl, and value.}
}
\value{
An entry data.frame with invalid probabilities all set to zero.
}
\description{
Sets invalid predictied probabilities (i.e. those summing to > 1.1 or < 0.9)
to zero. This function should be run prior to scoring an entry to ensure
no invalid probabilities are scored.
}
\examples{
valid_entry <- remove_invald(full_entry)
}
\seealso{
\code{\link{verify_entry}} \code{\link{score_entry}}
}

