% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_truth_1617}
\alias{state_truth_1617}
\title{Observed state ILI truth for 2016/2017 influenza season}
\format{A dta frame with 3992 rows and 4 variables:
\describe{
  \item{target}{character vector of targets}
  \item{location}{character vector of locations}
  \item{forecast_week}{numeric vector of MMWR week forecasted value is based on}
  \item{bin_start_incl}{numeric vector of observed truth}
}}
\usage{
state_truth_1617
}
\description{
A data.frame containing the observed state ILI forecasting truth for the
2016/2017 influenza season
}
\keyword{datasets}
