% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_exp_truth}
\alias{valid_exp_truth}
\title{Expanded national ILINet truth for 2015/2016 influenza season}
\format{A dta frame with 14212 rows and 4 variables:
\describe{
  \item{target}{character vector of targets}
  \item{location}{character vector of locations}
  \item{forecast_week}{numeric vector of MMWR week forecasted value is based on}
  \item{bin_start_incl}{numeric vector of observed truth}
}}
\usage{
valid_exp_truth
}
\description{
A data.frame containing the observed national ILI forecasting truth for the 
2015/2016 influenza season, expanded for one bin around week targets and 
five bins around percent targets
}
\keyword{datasets}
