% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_ILI}
\alias{valid_ILI}
\title{A valid ILINet submission}
\format{A dta frame with 363 rows and 3 variables:
\describe{
  \item{location}{character vector of locations}
  \item{week}{numeric vector of MMWR weeks}
  \item{ILI}{numeric vector of weighted ILI percentage}
}}
\usage{
valid_ILI
}
\description{
A data.frame containing a valid submission of national ILINet data
}
\keyword{datasets}
