% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_ILI_state}
\alias{valid_ILI_state}
\title{A valid state ILINet submission}
\format{A dta frame with 1860 rows and 3 variables:
\describe{
  \item{location}{character vector of locations}
  \item{week}{numeric vector of MMWR weeks}
  \item{ILI}{numeric vector of ILI percentage}
}}
\usage{
valid_ILI_state
}
\description{
A data.frame containing a valid submission of state ILI data
}
\keyword{datasets}
