% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_observe_hosp}
\alias{valid_observe_hosp}
\title{A valid FluSurv-NET submission}
\format{A dta frame with 180 rows and 3 variables:
\describe{
  \item{age_grp}{character vector of age groups}
  \item{week}{numeric vector of MMWR weeks}
  \item{weeklyrate}{numeric vector of weighted ILI percentage}
}}
\usage{
valid_observe_hosp
}
\description{
A data.frame containing a valid submission of FluSurv-Net data
}
\keyword{datasets}
