% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_agegrp.R
\name{verify_agegrp}
\alias{verify_agegrp}
\title{Verify the age groups in hospitalization forecasts}
\usage{
verify_agegrp(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
Invisibly returns \code{TRUE} if successful
}
\description{
Compares age groups to full entries and provides a warning if any
age groups are missing, an error if there are any extra age groups, and a
message to suggest forecasting for additional locations.
}
\keyword{internal}
