% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_colnames.R
\name{verify_colnames}
\alias{verify_colnames}
\title{Verify the column names of an entry}
\usage{
verify_colnames(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
Invisibly returns \code{TRUE} if successful
}
\description{
Compares column names to a valid entry and provides an error if any names
are missing and a warning if there are any extra names.
}
\keyword{internal}

