% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry.R
\name{verify_columns}
\alias{verify_columns}
\title{Check columns of an entry}
\usage{
verify_columns(entry)
}
\arguments{
\item{entry}{A data.frame}
}
\value{
Invisibly returns TRUE if the column names check out
}
\description{
Check to make sure the columns are location, target, type, unit,
bin_start_incl, bin_end_notincl, and value. If any of these are missing,
return an error indicating the column that is missing. If there are any
extra columns, return a warning indicating the extra columns.
}

