% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_hosp.R
\name{verify_hosp}
\alias{verify_hosp}
\title{Verify an observed hospitalizations file}
\usage{
verify_hosp(observe_hosp)
}
\arguments{
\item{observe_hosp}{A data.frame of observed FluSurv-NET values}
}
\value{
\code{NULL} or a descriptive error or warning message
}
\description{
Verifies user-supplied FluSurv-NET hospitalization data to create
observed truth
}
\examples{
verify_hosp(valid_observe_hosp)
}
