% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ILI.R
\name{verify_ILI_colnames}
\alias{verify_ILI_colnames}
\title{Verify supplied weighted ILI data columns}
\usage{
verify_ILI_colnames(weekILI)
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values.}
}
\value{
\code{NULL} or a descriptive error or warning message
}
\description{
Verifies columns in user-supplied wILI data
}
\keyword{internal}
