% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ILI.R
\name{verify_ILI_location}
\alias{verify_ILI_location}
\title{Verify supplied weighted ILI regions}
\usage{
verify_ILI_location(weekILI)
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values.}
}
\value{
\code{NULL} or a descriptive error or warning message
}
\description{
Verifies regions in user-supplied wILI data
}
\keyword{internal}
