% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_locations.R
\name{verify_locations}
\alias{verify_locations}
\title{Verify the locations}
\usage{
verify_locations(entry, challenge = "ilinet")
}
\arguments{
\item{entry}{An entry data.frame}

\item{challenge}{one of "ilinet" or "state_ili", indicating which
challenge the submission is for}
}
\value{
Invisibly returns \code{TRUE} if successful
}
\description{
Compares locations to minimal and full entries and provides an error if any
locations are missing, a warning if there are any extra locations, and a
message to suggest forecasting for additional locations.
}
\keyword{internal}
