% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_entry.R
\name{verify_structure}
\alias{verify_structure}
\title{Verify the entry structure}
\usage{
verify_structure(entry, valid_entry)
}
\arguments{
\item{entry}{An entry data.frame}

\item{valid_entry}{A valid entry data.frame}
}
\value{
NULL or a character vector of error messages
}
\description{
Verify the entry structure
}

