% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_types.R
\name{verify_types}
\alias{verify_types}
\title{Verify types are correct}
\usage{
verify_types(entry, challenge = "ilinet")
}
\arguments{
\item{entry}{An entry data.frame}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating which
challenge the submission is for}
}
\value{
Invisibly returns TRUE if successful
}
\description{
The necessary types depend on the type, so this will verify types are correct
for all types
}
\examples{
verify_types(minimal_entry)
}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}
