% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_entry.R
\name{write_valid_entry}
\alias{write_valid_entry}
\title{Write a valid entry to a csv file}
\usage{
write_valid_entry(path)
}
\arguments{
\item{path}{A character string specifying a directory}
}
\value{
Invisibly returns TRUE if successful
}
\description{
This will write a complete valid entry to a csv file so participants can
see the structure of the file. The file will be called
}
\seealso{
\code{\link{write_entry}}
}
