% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry.R
\name{arrange_entry}
\alias{arrange_entry}
\title{Arrange an entry for consistency}
\usage{
arrange_entry(entry)
}
\arguments{
\item{entry}{A data.frame}
}
\value{
An arranged data.frame
}
\description{
Arrange an entry for consistency
}
\keyword{internal}
