% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasonal.R
\name{create_peak}
\alias{create_peak}
\title{Creates observed truth for seasonal peak values}
\usage{
create_peak(weekILI, location, challenge = "ilinet")
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values (default NULL). 
Must contain columns location (or age_grp), week, and wILI.}

\item{location}{A character string specifying the target location or age group}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating which
challenge the submission is for (default \code{"ilinet"})}
}
\value{
A data.frame with columns location (or age_grp), target, and bin_start_incl
}
\description{
Determines observed true values for peak week and peak weighted ILINet
percentage for an influenza season
}
\keyword{internal}
