% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_truth.R
\name{create_truth}
\alias{create_truth}
\title{Creates observed truth to score entry against}
\usage{
create_truth(fluview = TRUE, year = NULL, weekILI = NULL)
}
\arguments{
\item{fluview}{A logical value (default \code{TRUE}) indicating whether
to download ILINet from Fluview}

\item{year}{Calendar year during which the flu season of interest begins. 
For the 2015/2016 flu season, \code{year = 2015}}

\item{weekILI}{A data.frame of weighted ILI values (default \code{NULL}).
Must contain columns location, week, and wILI. Must be \code{NULL} when 
downloading data using \code{fluview = TRUE}. Required if \code{fluview 
= FALSE.}}
}
\value{
A data.frame with columns location, target, and bin_start_incl
}
\description{
Determines observed true values for each target
}
\examples{
truth <- create_truth(year = 2015)
truth <- create_truth(fluview = TRUE, year = 2015)
truth <- create_truth(fluview = FALSE, weekILI = valid_ILI)

\dontrun{
truth <- create_truth(weekILI = valid_ILI)
truth <- create_truth(fluview = FALSE)
}

}

