% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_truth.R
\name{create_truth}
\alias{create_truth}
\title{Creates observed truth to score entry against}
\usage{
create_truth(fluview = TRUE, year = NULL, weekILI = NULL,
  challenge = "ilinet", start_wk = NULL, end_wk = NULL)
}
\arguments{
\item{fluview}{A logical value (default \code{TRUE}) indicating whether to
download ILINet from Fluview.}

\item{year}{Calendar year during which the flu season of interest begins. For
the 2015/2016 flu season, \code{year = 2015}. Required whether downloading
data from fluview or providing observed data}

\item{weekILI}{A data.frame of observed values (default \code{NULL}). Must be
  \code{NULL} when downloading data using \code{fluview = TRUE}. Required
  if \code{fluview = FALSE}.

  For \code{challege = "ilinet"} or \code{challenge = "state_ili"}, must
  contain columns location, week, and wILI. For \code{challenge = "hospital"},
  must contain columns age_grp, week, and weeklyrate.}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating which
challenge the submission is for (default \code{"ilinet"})}

\item{start_wk}{MMWR week that defines the start of the challenge. Default value 
provided for years greater than \code{2014} that aligns with start week of 
CDC challenge. Otherwise defaults to MMWR week 42 for
\code{challenge == "ilinet"} and \code{challenge == "state"},
or MMWR week 48 for \code{challenge == "hospital"}}

\item{end_wk}{MMWR week that defines the end of the challenge. Default value 
provided for years greater than \code{2014} that aligns with end week of 
CDC challenge. Otherwise defaults to MMWR week 18 for
\code{challenge == "ilinet"} and \code{challenge == "state"},
or MMWR week 17 for \code{challenge == "hospital"}}
}
\value{
A data.frame with columns location, target, and bin_start_incl or
  columns age_grp, target, and bin_start_incl if \code{challenge = "hospital"}
}
\description{
Determines observed true values for each target
}
\examples{
truth <- create_truth(year = 2015)
truth <- create_truth(fluview = TRUE, year = 2015, challenge = "ilinet")
truth <- create_truth(fluview = FALSE, weekILI = valid_ILI)
truth <- create_truth(fluview = TRUE, year = 2015, challenge = "hospital")

\dontrun{
truth <- create_truth(weekILI = valid_ILI)
truth <- create_truth(fluview = FALSE)
}

}
