% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_point_forecasts.R
\name{generate_point_forecasts}
\alias{generate_point_forecasts}
\title{Generate point forecasts for all locations and targets}
\usage{
generate_point_forecasts(entry, method = c("Median", "Expected Value",
  "Mode"), challenge = "ilinet")
}
\arguments{
\item{entry}{An entry data.frame}

\item{method}{The method to be used to generate the point forecasts. 
\code{"Median"} (the default) uses the median value, \code{"Expected
Value"} generates the expected value from the provided probabilities,and
\code{"Mode"} returns the individual bin with the largest probability}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating
which challenge the submission is for (default \code{"ilinet"}).}
}
\value{
A data.frame of point forecasts for all locations and targets.
}
\description{
Generate point forecasts for all locations and targets
}
\seealso{
\code{\link{generate_point_forecast}}, \code{\link{verify_entry}}
}
