% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_probs.R
\name{normalize_probs}
\alias{normalize_probs}
\title{Normalize probabilistic predictions}
\usage{
normalize_probs(entry, ignore_invalid = FALSE, challenge = "ilinet")
}
\arguments{
\item{entry}{A verified entry data.frame with columns location, target, type, unit, 
bin_start_incl, bin_end_incl, and value.}

\item{ignore_invalid}{logical; if \code{FALSE} all location/target combinations
have probabilities normalized: if \code{TRUE} location/target
combinations with probabilities summing to < 0.9 or > 1.1 are not
normalized. Should be \code{TRUE} and run in combination with 
\code{\link{remove_invalid}} if run on an entry prior to scoring.}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating
which challenge the submission is for (default \code{"ilinet"}).}
}
\value{
An entry data.frame with normalized probabilities
}
\description{
Normalizes all probabilistic predictions so that the cumulative probability 
for a target sums to 1.
}
\examples{
norm_entry <- normalize_probs(full_entry)
}
\seealso{
\code{\link{verify_entry}} \code{\link{remove_invalid}}
}
