% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_bins.R
\name{verify_bins}
\alias{verify_bins}
\title{Verify bins are correct}
\usage{
verify_bins(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
Invisibly returns TRUE if successful
}
\description{
The necessary bins depend on the target, so this will verify bins are correct
for all targets
}
\examples{
verify_bins(minimal_entry)
}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}

