% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_colnames.R
\name{verify_colnames}
\alias{verify_colnames}
\title{Verify the column names of an entry}
\usage{
verify_colnames(entry, check_week = T)
}
\arguments{
\item{entry}{An entry data.frame}

\item{check_week}{A logical value (default `TRUE`) indicating whether to check
for the column forecast_week. Should be `TRUE` if evaluating entry prior to 
scoring, can be `FALSE` if evaluating entry prior to writing to disk.}
}
\value{
Invisibly returns \code{TRUE} if successful
}
\description{
Compares column names to a valid entry and provides an error if any required names
are missing and a warning if there are any extra names.
}
\keyword{internal}
