% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_entry.R
\name{verify_entry}
\alias{verify_entry}
\title{Verifies a csv submission file}
\usage{
verify_entry(file)
}
\arguments{
\item{file}{A csv file}
}
\value{
TRUE if the file is compatible and FALSE if it is not.
}
\description{
This function will run through a variety of tests to check a submission file.
It will return TRUE if the entry is valid and FALSE if not. In both cases, it might
print warnings.
}
\examples{
foo <- system.file("extdata", "valid-test.csv", package="FluSight")
verify_entry(foo)
}

