% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_entry.R
\name{verify_entry}
\alias{verify_entry}
\title{Verify entry stored as an R data.frame}
\usage{
verify_entry(entry)
}
\arguments{
\item{entry}{A data.frame}
}
\value{
TRUE or a character vector of errors
}
\description{
Verify entry stored as an R data.frame
}
\examples{
file <- system.file("extdata", "valid-test.csv", package="FluSight")
entry <- read.csv(file, stringsAsFactors = FALSE)
verify_entry(entry) # TRUE
}
\seealso{
verify_entry_file
}

