% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_entry.R
\name{verify_entry}
\alias{verify_entry}
\title{Verify entry stored as an R data.frame}
\usage{
verify_entry(entry)
}
\arguments{
\item{entry}{A data.frame}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Verify entry stored as an R data.frame
}
\examples{
verify_entry(minimal_entry)
verify_entry(full_entry)
}
\seealso{
\code{\link{verify_entry_file}}
}

