% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ILI.R
\name{verify_ILI}
\alias{verify_ILI}
\title{Verify supplied weighted ILI data}
\usage{
verify_ILI(weekILI, challenge = "ilinet")
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values.}

\item{challenge}{one of "ilinet" or "state_ili", indicating which
challenge the submission is for (default \code{"ilinet"})}
}
\value{
\code{NULL} or a descriptive error or warning message
}
\description{
Verifies user-supplied wILI data to be used to generate forecasting targets
}
\examples{
verify_ILI(valid_ILI)

}
