% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ILI.R
\name{verify_ILI}
\alias{verify_ILI}
\title{Verify supplied weighted ILI data}
\usage{
verify_ILI(weekILI)
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values.}
}
\value{
NULL or a descriptive error or warning message
}
\description{
Verifies user-supplied wILI data to be used to generate forecasting targets
}

