% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ILI.R
\name{verify_ILI_location}
\alias{verify_ILI_location}
\title{Verify supplied weighted ILI regions}
\usage{
verify_ILI_location(weekILI, challenge = "ilinet")
}
\arguments{
\item{weekILI}{A data.frame of weighted ILI values.}

\item{challenge}{one of "ilinet" or "state_ili", indicating which
challenge the submission is for (default \code{"ilinet"})}
}
\value{
\code{NULL} or a descriptive error or warning message
}
\description{
Verifies regions in user-supplied wILI data
}
\keyword{internal}
