% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_entry.R
\name{verify_point}
\alias{verify_point}
\title{Verify validity of point predictions}
\usage{
verify_point(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
NULL or a descriptive warning/error message
}
\description{
Verify validity of point predictions
}

