% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_probabilities.R
\name{verify_probabilities}
\alias{verify_probabilities}
\title{Verify the entry probabilities}
\usage{
verify_probabilities(entry, challenge = "ilinet")
}
\arguments{
\item{entry}{An entry data.frame}

\item{challenge}{one of "ilinet", "hospital" or "state_ili", indicating which
challenge the submission is for}
}
\value{
Invisibly returns \code{TRUE} or a descriptive error message
}
\description{
Verify the entry probabilities
}
\examples{
verify_probabilities(minimal_entry)
verify_probabilities(full_entry_hosp, challenge = "hospital")

}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}
