% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_probabilities.R
\name{verify_probabilities}
\alias{verify_probabilities}
\title{Verify the entry probabilities}
\usage{
verify_probabilities(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
Invisibly returns \code{TRUE} or a descriptive error message
}
\description{
Verify the entry probabilities
}
\examples{
verify_probabilities(minimal_entry)
verify_probabilities(full_entry_hosp)

}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}
