% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_entry.R
\name{verify_probabilities}
\alias{verify_probabilities}
\title{Verify the entry probabilities}
\usage{
verify_probabilities(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
NULL or a character vector of error messages
}
\description{
Verify the entry probabilities
}

