% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_targets.R
\name{verify_targets}
\alias{verify_targets}
\title{Verify targets are correct}
\usage{
verify_targets(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
Invisibly returns TRUE if successful
}
\description{
The necessary targets depend on the target, so this will verify targets are correct
for all targets
}
\examples{
verify_targets(minimal_entry)
}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}
