% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_units.R
\name{verify_units}
\alias{verify_units}
\title{Verify units are correct}
\usage{
verify_units(entry)
}
\arguments{
\item{entry}{An entry data.frame}
}
\value{
Invisibly returns TRUE if successful
}
\description{
The necessary units depend on the target, so this will verify units are correct
for all units
}
\examples{
verify_units(minimal_entry)
}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}
