% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_entry.R
\name{write_entry}
\alias{write_entry}
\title{Write an entry to a csv file}
\usage{
write_entry(entry, path, team_name, week = NULL)
}
\arguments{
\item{entry}{An entry data.frame}

\item{path}{A character string specifying a directory for the file}

\item{team_name}{A character string specifying the team name}

\item{week}{A scalar numeric indicating the week number or NULL (default).
If NULL, the week will automatically be determined from the current date.}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Verify an entry and then write it to a csv file
}
\seealso{
\code{\link{write_valid_entry}}, \code{\link{read_entry}}
}

